requirements.txt
streamlit
google-generativeai
python-dotenv
googlesearch-python
Pillow